backend/requirements.txt
python-docx==0.8.11
pillow==10.0.1
